+++ lib/json_api_client/helpers/dirty.rb
@@ -18,6 +18,10 @@ module JsonApiClient
         @changed_attributes = ActiveSupport::HashWithIndifferentAccess.new
       end
 
+      def forget_change!(attr)
+        @changed_attributes.delete(attr.to_s)
+      end
+
       def set_all_attributes_dirty
         attributes.each do |k, v|
           set_attribute_was(k, v)